requirements.txt
snowflake-connector-python>=3.3.0
streamlit[snowflake]>=1.54.0
altair>=5.5.0
pandas>=2.2.3
